ui-tests/browseruse-self-healing/requirements.txt
browser-use
langchain-anthropic
pydantic>=2.0
pytest
pytest-asyncio
python-dotenv
